# bookstore-front/src/app/service/model/Book.ts
/**
 * BookStore APIs
 * Book Store APIs
 *
 * OpenAPI spec version: 1.0.0
 * Contact: jinbaek7@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * Book resource representation
 */
export class Book {
    /**
     * Identifier
     */
    "id"?: number;
    /**
     * Title of the book
     */
    "title": string;
    /**
     * Description of the book
     */
    "description"?: string;
    /**
     * Cost of the book
     */
    "unitCost"?: number;
    /**
     * ISBN number
     */
    "isbn": string;
    /**
     * Publication date of the book
     */
    "publicationDate"?: Date;
    /**
     * Number of pages
     */
    "nbOfPages"?: number;
    /**
     * URL of the book cover
     */
    "imageUrl"?: string;
    /**
     * Language that the book was written in
     */
    "language"?: Book.LanguageEnum;

}

export namespace Book {
    export enum LanguageEnum {
        ENGLISH = <any> 'ENGLISH',
        FRENCH = <any> 'FRENCH',
        SPANISH = <any> 'SPANISH',
        PORTUGUESE = <any> 'PORTUGUESE',
        ITALIAN = <any> 'ITALIAN',
        FINISH = <any> 'FINISH',
        GERMAN = <any> 'GERMAN',
        DEUTSCH = <any> 'DEUTSCH',
        RUSSIAN = <any> 'RUSSIAN'
    }
}